backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
motor==3.4.0
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.3
pyspark==3.5.1
torch==2.2.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
web3==6.18.0
eth-account==0.10.0
python-dotenv==1.0.1
httpx==0.27.0
kafka-python==2.0.2
structlog==24.1.0
pytest==8.1.1
pytest-asyncio==0.23.6
